feat: Atualização da modelagem da avaliação e implementação do usecase de avaliação referente aos dtos e as entidades.

+++ src/main/java/com/fiap/pos_tech/agendamento_servicos/infrastructure/persistence/entity/AvaliacaoEntity.java
@@ -27,7 +27,7 @@ public class AvaliacaoEntity {
     @JoinColumn(name = "agendamento_id", nullable = false)
     private AgendamentoEntity agendamento;
 
-    private Long nota;
+    private Integer nota;
 
     private String comentario;
